+++ R/functions-Spectrum1.R
@@ -1,9 +1,10 @@
 show_Spectrum1 <- function(spectrum) {
-  cat("Object of class \"", class(spectrum), "\"\n", sep="")
-  cat(" Retention time:", formatRt(spectrum@rt), "\n")
-  cat(" MSn level:", spectrum@msLevel, "\n")
-  cat(" Total ion count:", spectrum@peaksCount, "\n")
-  cat(" Polarity:", spectrum@polarity, "\n")
+    cat("Object of class \"", class(spectrum), "\"\n", sep="")
+    if (length(spectrum@rt))
+        cat(" Retention time:", formatRt(spectrum@rt), "\n")
+    cat(" MSn level:", spectrum@msLevel, "\n")
+    cat(" Total ion count:", spectrum@peaksCount, "\n")
+    cat(" Polarity:", spectrum@polarity, "\n")
 }
 
 ############################################################