+++ NEWS.md
@@ -4,6 +4,7 @@
 
 - Re-implement `impute,MSnSet-method` with
   `MsCoreUtils::impute_matrix()`.
+- Illustrate `ReporterIons` class using TMT16.
 
 ## Changes in 2.17.3
 - Add `normalize,Chromatogram` method.